requirements.txt
google-generativeai
Pillow
markdown2
openai